Made refactoring and removed some classes
